removed more old code from various scripts. duped WorkerBrainStrategy to make tweaks to new one. added PoiHealFriendliesStrategy to replace healing at food shack location.
spent 3 hrs trying to set up IInteractables interface to work in a similar way to how EntityActions are. where i only need an EntityActionStrategy then pass in a class that uses the interface.
cant figure it out/find a good way of getting it to work how i intend it to work.