added AiDirector script and class: moved all code in GM script into this class + made slight changes to events and funcs etc...
spent 2hrs coming up with a more indepth solution for allowing entities to call and answer calls for help and partially implementing it.

added CallForHelpStrategy class to EntityActonStrategies script:
added logic to get closest entities from position based on team, max search limit and number of entities it should try to find.

TODO:
set up GOAP beliefs, actions and goals to allow it to call for help and answer call for help then tweak it.