spent 3hrs reworking existing code to get request and answer request help calls set up.
added new sensor type friendly sensor + code to support it in various scripts and tested it works as a sensor and detects the correct entities.

updated CallForHelpStrategy class:
reworked existing code into a much simpler version that works off of friendly sensors list of TargetData. + checks and calls to various functions to decide if an entity can be requested for help or not.

updated EntityBrain class:
added a timer to limit the amount of requests an entity can make + a function to simplify checking if an entity does need help and a function to check if an entity can answer a request help call.

updated CombatBrainStrategy class:
added beliefs, actions and goals for request help and tested it works with existing GOAP behaviour.

TODO:
finish the answer requst help side of GOAP behaviour as its half implemented at the moment and should only require new beliefs, a move action and goal.