spent 3hrs fully reworking EntitySensors and trying to get them to work with grabbing a target from a sensors list based on what i need in the entity taget belief.
also made a lot of tweaks to brains themselves.

current issue:
old sensors could force entity brain to recalc beliefs actions and goals base on OnTargetChanged event.
this allows move to target to be accurate to targets position as it changes + force out of current goals and actions when a new target is found/lost + includ checks to only force the recalc if target is now a different target.

spent 30mins thinking of ways to fix this issue.

TODO:
implement following solution to see if it works + anything else i can think of.
storing each TargetData i get on the sensor as attackOneTarget, attackTwoTarget, friendlyPoi etc...
have the beliefs just read that and call all the funcs in UpdateSensorLogic() func, invoke OnTargetChanged event with new target change type enum.